0.11 Catalog sample uses the framework it demonstrates
ProductDetail rendered "not found" as inline Text while Products used
ContentState.Empty for the same shape of result — a success that found
nothing. It now uses ContentState.Empty too, under its own content id so
onSystemEvent tells the "Go back" action from a retry of a failed load and
answers it with UiCommand.NavigateBack. ProductDetailState.product stops
being nullable as a result.

Product.price becomes a Long of minor units and asPrice() formats it with
NumberFormat.getCurrencyInstance(locale), taking the scale from the currency,
rather than "$%.2f".

Gate: doctor 17/17, build green.

+++ feature/catalog/domain/src/main/kotlin/com/example/androidproject1/feature/catalog/domain/Product.kt
@@ -1,9 +1,13 @@
 package com.example.androidproject1.feature.catalog.domain
 
+/**
+ * @property price in minor units — 450 is 4.50. Integer money cannot drift the way a `Double` can,
+ * and it leaves the currency and its formatting to the layer that knows the user's locale.
+ */
 data class Product(
     val id: String,
     val categoryId: String,
     val name: String,
-    val price: Double,
+    val price: Long,
     val description: String,
 )
